Update[3/5]: Call notifier chain only when temps updated
update_temps is used to hold the updated temperature and when current temprature changes the notifier chain is triggered due to mismatch between cur and updated temperatures and updated_temps is reset with the value of current temps.
this reduces the amount of times the notifier chain is called and saves unnessary JUMP and stack operations consuming cpu cycles.

Signed-off-by: Tejas Udupa <tejasudupa1285@gmail.com>