Cleanup sysfs group after exiting the governor
We stumble on kernel panic after anon_vma_page alloc fails to allocte page as
we re-allocate each time we restart the governor. this leads to memory not being
given back to kernel after use and thus kernel panic after multiple restarts
of the governor. it takes a while for the panic to happen and it may not happen
right after governor is restarted it may also happen when governor not in use
so it depends on when that anon_vma_page is called.

Signed-off-by: Tejas Udupa <tejasudupa1285@gmail.com>